// frontend/src/apis/models/ChatPlatform.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * TellMom API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ChatPlatform = {
    Roblox: 'roblox',
    Discord: 'discord',
    Minecraft: 'minecraft'
} as const;
export type ChatPlatform = typeof ChatPlatform[keyof typeof ChatPlatform];


export function instanceOfChatPlatform(value: any): boolean {
    for (const key in ChatPlatform) {
        if (Object.prototype.hasOwnProperty.call(ChatPlatform, key)) {
            if (ChatPlatform[key as keyof typeof ChatPlatform] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ChatPlatformFromJSON(json: any): ChatPlatform {
    return ChatPlatformFromJSONTyped(json, false);
}

export function ChatPlatformFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChatPlatform {
    return json as ChatPlatform;
}

export function ChatPlatformToJSON(value?: ChatPlatform | null): any {
    return value as any;
}

export function ChatPlatformToJSONTyped(value: any, ignoreDiscriminator: boolean): ChatPlatform {
    return value as ChatPlatform;
}