// frontend/src/apis/models/ParentResponse.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * TellMom API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ParentResponse
 */
export interface ParentResponse {
    /**
     * 
     * @type {number}
     * @memberof ParentResponse
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof ParentResponse
     */
    email: string;
}

/**
 * Check if a given object implements the ParentResponse interface.
 */
export function instanceOfParentResponse(value: object): value is ParentResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    return true;
}

export function ParentResponseFromJSON(json: any): ParentResponse {
    return ParentResponseFromJSONTyped(json, false);
}

export function ParentResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ParentResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'email': json['email'],
    };
}

export function ParentResponseToJSON(json: any): ParentResponse {
    return ParentResponseToJSONTyped(json, false);
}

export function ParentResponseToJSONTyped(value?: ParentResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'email': value['email'],
    };
}